Improve random identifier in random name helper
The PrefixedRandomName helper generates names of the form "CATS-1-ORG-6e0b3df3-c397-4643-4", where the string of hexadecimal digits comes from truncating a v4 UUID at 20 characters. This change simplifies the helper to make its random identifier simply a string of 16 hexadecimal digits, so that random names are instead of the form "CATS-1-ORG-6e0b3df3c3974643". This consolidated identifier will be easier to find in aggregated test and component logs and will not be confused with other UUIDs internal to Cloud Foundry.

This change also removes the dependency on the third-party UUID library in favor of the equivalent calls to the standard library.